+++ src/ShellKrypt.Core/Authenticator/IOtpAuthUriParser.cs
@@ -1,0 +1,6 @@
+namespace ShellKrypt.Core.Authenticator;
+
+public interface IOtpAuthUriParser
+{
+    ParsedOtpAuthSecret Parse(string otpauthUri);
+}